separate layers and start conv and some weird errors
