+++ matrix.h
@@ -42,3 +42,4 @@ Matrix *dot(Matrix *a, Matrix *b);
 Matrix *copy_matrix(Matrix *matrix);
 float sum(Matrix *matrix);
 Matrix *create_matrix_from_array(int rows, int cols, float *data);
+Matrix *pad(Matrix *matrix, int pad_size);
