+++ nn.h
@@ -3,7 +3,6 @@
 //
 // Created by plusleft on 10/25/2023.
 //
-#include <stddef.h>
 #include "matrix.h"
 #include "data.h"
 #include "layers.h"
@@ -23,4 +22,3 @@ void destroy_network(Network *network);
 // maybe return final loss?
 void train_dataset(Network *network, Dataset *dataset, int epochs, float learning_rate);
 float train_input(Network *network, Matrix *input, Matrix *expected, float learning_rate);
-void initialize_weights_xavier(Network *network);
